# `pivot_wider()` performance