Feel free to close this PR and just use it as something to refer to later on.

Using the new `vec_group_id()` with `vec_unique_loc()` we can get some nice performance improvements over the current `pivot_wider()` implementation.

Using https://github.com/tidyverse/tidyr/issues/775 as an example

``` r
library(tidyr)
library(dplyr, warn.conflicts = FALSE)
library(reshape2)

mydf <- expand_grid(
  case = sprintf("%03d", seq(1, 4000)),
  year = seq(1900, 2000),
  name = c("x", "y", "z")
) %>%
  mutate(value = rnorm(nrow(.)))
```

This is on current dev tidyr with current dev vctrs. It is already slightly faster than what is in the provided example. Maybe we made some vctrs improvements, not sure.

``` r
bench::mark(
  pivot = pivot_wider(mydf, names_from = "name", values_from = "value"),
  spread = spread(mydf, name, value),
  dcast = dcast(mydf, case + year ~ name),
  iterations = 50
)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 3 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 pivot         815ms    881ms      1.14     115MB     2.45
#> 2 spread        430ms    504ms      2.01     431MB    11.1 
#> 3 dcast         407ms    485ms      2.07     457MB     8.82
```

Now with this PR

```r
bench::mark(
  pivot = pivot_wider(mydf, names_from = "name", values_from = "value"),
  spread = spread(mydf, name, value),
  dcast = dcast(mydf, case + year ~ name),
  iterations = 50
)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 3 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 pivot         535ms    617ms      1.59     117MB     3.41
#> 2 spread        419ms    512ms      1.97     431MB     8.60
#> 3 dcast         417ms    458ms      2.11     457MB     7.42
```