@krlmlr 2019-11-11
## Before

``` r
library(tidyverse)

set.seed(20191111)

N <- 100

data <-
  crossing(
    key1 = factor(letters),
    key2 = factor(letters),
    key3 = Sys.Date() - seq_len(N),
    nesting(name = factor(LETTERS), value = runif(26))
  )

system.time(
  pivot_wider(data, names_from = name)
)
#>    user  system elapsed 
#>   1.359   0.056   1.415
```

<sup>Created on 2019-11-11 by the [reprex package](https://reprex.tidyverse.org) (v0.3.0)</sup>

## After

``` r
system.time(
  pivot_wider(data, names_from = name)
)
#>    user  system elapsed 
#>   0.818   0.032   0.850
```

<sup>Created on 2019-11-11 by the [reprex package](https://reprex.tidyverse.org) (v0.3.0)</sup>