@krlmlr 2019-11-11
Note that with the CRAN version of vctrs, the performance is abysmal. This is with `N <- 10`:

``` r
system.time(
  pivot_wider(data, names_from = name)
)
#>    user  system elapsed 
#>   9.868   0.004   9.873
```

<sup>Created on 2019-11-11 by the [reprex package](https://reprex.tidyverse.org) (v0.3.0)</sup>